BACKLOG-38036 - Moved classes to fix circular dependency problem between scheduler and extensions
